+++ src/fem/simple_mesh.h
@@ -9,7 +9,7 @@
 #include "../binding_core.h"
 
 
-void simple_gmesh(std::string filename);
+void simple_gmesh(const std::string& filename);
 
 void gmsh_module(nb::module_ &m);
