+++ src/visit/visit_py_wrap.cpp
@@ -3,7 +3,8 @@
 //
 #include "../geom/Mesh.h"
 #include "tess_helpers.h"
-#include "tess_py_wrap.h"
+#include "visit_py_wrap.h"
+#include "TessellateFactory.h"
 
 void tess_helper_module(nb::module_ &m) {
     m.def("get_box_mesh", &get_box_mesh, "box_origin"_a, "box_dims"_a, "Write a box to a step file");
@@ -23,4 +24,18 @@ void tess_helper_module(nb::module_ &m) {
             .def_rw("g", &Color::g)
             .def_rw("b", &Color::b)
             .def_rw("a", &Color::a);
+}
+
+void tess_module(nb::module_ &m) {
+    nb::enum_<TessellationAlgorithm>(m, "TessellationAlgorithm")
+            .value("OCCT_DEFAULT", TessellationAlgorithm::OCCT_DEFAULT,
+                   "Converts to OCC TopoDS_Shape and uses BRepMesh_IncrementalMesh")
+            .value("CGAL_DEFAULT", TessellationAlgorithm::CGAL_DEFAULT, "Uses CGAL's polyhedron_3");
+
+    nb::class_<TessellateFactory>(m, "TessellateFactory")
+            .def(nb::init<>())
+            .def(nb::init<const TessellationAlgorithm &>())
+            .def("tessellate", &TessellateFactory::tessellate)
+            .def_rw("algorithm", &TessellateFactory::algorithm, "Tessellation algorithm");
+
 }
